fix prelistingid4

+++ app/api/seed/route.ts
@@ -1,36 +1,30 @@
 import prisma from "@/lib/prisma";
 import { v4 as uuidv4 } from 'uuid';
 import { NextResponse } from "next/server"; 
+import { auth } from "@/auth";
 
 export async function GET(request:Request){
-
+  const session = await auth();
+  if (!session?.user || !session.user.id){
+    return new Response('Unauthorized', { status: 401 });
+  }
     await prisma.prelisting.deleteMany()
     await prisma.form.deleteMany()
     const prelisting= await prisma.prelisting.create({
-        data: { id:uuidv4(),title:'Messi', description:'Argento'},
-        // {id:uuidv4(),title:'Suarez', description:'Uruguayo'},
-        // {id:uuidv4(),title:"busquet", description:'Espanol'},
-        // {id:uuidv4(),title:"Alba", description:"Gallego"}
+        data: { id:uuidv4(),userId: session.user.id ,title:'Messi', description:'Argento'}, 
+         
+   
     
 })
 console.log("%c Line:10 🥪 prelisting", "color:#6ec1c2", prelisting);
   
   const createdForms = await prisma.form.create({
     data: {
-        asesor: 'Janet', email: 'janet@gmail.com', celular: 2234, fechadenacimiento: '12-03-1990', porquevende: "porque se le canta", proprietario: "lana",
+        asesor: 'Janet', email: 'janet@gmail.com', celular: 2234, fechadenacimiento: '12-03-1990', whysell: "porque se le canta", proprietario: "lana",
         prelisting: {
-            connect: { id: prelisting.id } // Connect to an existing prelisting
+            connect: { id: prelisting.id } 
           }
-    // },{
-    //     asesor: 'Sebas', email: 'sebas@gmail.com', celular: 2234, fechadenacimiento: '12-03-1990', porquevende: "porque se le canta", proprietario: "lana",
-    //     prelistingId:''
-    // },{
-    //     asesor: 'Ana', email: 'ana@gmail.com', celular: 2234, fechadenacimiento: '12-03-1990', porquevende: "porque se le canta", proprietario: "lana",
-    //     prelistingId:''
-    // },{
-    //     asesor: 'Jose', email: 'jose@gmail.com', celular: 2234, fechadenacimiento: '12-03-1990', porquevende: "porque se le canta", proprietario: "lana",
-    //     prelistingId:''
-    // }]
+   
   }});
   
   console.log("Created forms:", createdForms);
